feat: dffuse and specular lighting and shadows

+++ src/material.rs
@@ -3,12 +3,12 @@ use crate::color::Color;
 
 pub struct Material {
     pub base_color: Color,
-    pub specular_amount: f32,
-    pub shininess: f32
+    pub specular_amount: f64,
+    pub shininess: f64
 }
 
 impl Material {
-    pub fn new(base_color: Color, specular: f32, shininess: f32) -> Material {
+    pub fn new(base_color: Color, specular: f64, shininess: f64) -> Material {
         Material {
             base_color,
             specular_amount: specular,
